minor; addressing lint suggestions
