+++ src/components/TimeMachine.tsx
@@ -1,5 +1,5 @@
 import { useRef, useEffect, useState, useCallback } from "react";
-import { useGameState } from "@/context/game-state/provider";
+import { useGameState } from "@/context/game-state/hook";
 import { useControls, button } from "leva";
 import { SnapshotFrom } from "xstate";
 import { TurnMachine } from "@/state/machines/turn";
