+++ src/context/game-state/hook.tsx
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { stateContext } from "./provider";
+
+export const useGameState = () => {
+  const context = useContext(stateContext);
+  if (!context) {
+    throw new Error("useGameState must be used within a GameStateProvider");
+  }
+  return context;
+};
